refactor(toolkit): move graph db to toolkit
