+++ .github/pull_request_template.md
@@ -27,21 +27,24 @@ And make sure that the title of your pull request follows the following format:
 **Scope**
 <!-- Which module does your pull request mainly modify? Select `none` when undetermined. -->
 
-- [ ] `web`: (web front-end module)
-- [ ] `server`: (web server module)
-- [ ] `dal`: (data access layer)
-- [ ] `sdk`: (sdk module)
-- [ ] `planner`: (planner/leader module)
-- [ ] `workflow`: (workflow module)
-- [ ] `knowledge`: (knowledge module)
-- [ ] `toolkit`: (toolkit module)
-- [ ] `reasoner`: (reasoner module)
-- [ ] `memory`: (memory module)
-- [ ] `model`: (model/LLM module)
-- [ ] `env`: (env module)
-- [ ] `resource`: (resource module)
-- [ ] `tracer`: (tracer module)
-- [ ] `plugin`: (plugin module)
+- [ ] `app`: (**Application Layer**)
+  - [ ] `web`: (web front-end module)
+  - [ ] `server`: (web server module)
+  - [ ] `dal`: (data access layer)
+  - [ ] `sdk`: (sdk module)
+- [ ] `agent`: (**Agent Layer**)
+  - [ ] `model`: (model/LLM module)
+  - [ ] `reasoner`: (reasoner module)
+  - [ ] `planner`: (planner/leader module)
+  - [ ] `workflow`: (workflow module)
+  - [ ] `memory`: (memory module)
+  - [ ] `knowledge`: (knowledge module)
+  - [ ] `env`: (env module)
+  - [ ] `toolkit`: (toolkit module)
+- [ ] `system`: (**System Layer**)
+  - [ ] `resource`: (resource module)
+  - [ ] `tracer`: (tracer module)
+  - [ ] `plugin`: (plugin module)
 - [ ] `none`: (no module)
 
 ### Description
