+++ app/core/toolkit/graph_db/graph_db_factory.py
@@ -1,6 +1,6 @@
 from app.core.common.type import GraphDbType
-from app.core.graph_db.graph_db import GraphDb
-from app.core.graph_db.graph_db_config import GraphDbConfig, Neo4jDbConfig, TuGraphDbConfig
+from app.core.toolkit.graph_db.graph_db import GraphDb
+from app.core.model.graph_db_config import GraphDbConfig, Neo4jDbConfig, TuGraphDbConfig
 
 
 class GraphDbFactory:
